Add new Line Switch function
The "line_switch" bug function was added to "bf.function_list". It chooses a random line from the file and switches
place with the line before it. If either of the lines are empty, the functions runs again.
If the chosen line is the first line in the file, it switches place with the line after it.