Added rough draft of if_switch()
This function is supposed to make an if statement
into a regular statement and a non-if statement
into an if-statement. It is not working quite as intended
and requires further review. Perhaps it needs more conditionals
to ensure it doesn't apply to the wrong lines, but
it is already quite lengthy.

Perhaps each of these abilites will be split up into their own
respective functions.

In general, if the purpose of this program is to introduce bugs
and this function has some unintended consequences,
is it working correctly? Ideally, this program would introduce
logical errors instead of syntax errors which are far easier
to debug since the program won't run.