Modified random_line() function.
random_line() function was given an option argument
of opt_arg = 'line_list'. By default, this function
will return the random_line_index as well as a list of
the characters in that line. However if you specify an additional
argument such as random_line(filelist, 'no_line_list') as in the
case of the if_switch() function, then the full line is returned
as a string.

In this way, the function is more versatile and useful and can be
used in more functions. We can abstract the conditionals to the
random_line() function and away from the bug functions themselves.

By applying this to the if_switch() function, the problem of it being
applied to comments is now resolved.

A review of the other function is warranted so they can all use
the random_line function appropriately.