Added optional argument to alter_file_name()
By default, a file passed to alter_file_name() will retain
its original file extension. However, by passing 'text' to it,
it will add a ".txt" to the end of the file making it into a
simple text file.

This was done to make it easier to review the Diff Answer Key
by double cliking on it. Prior to this, double clicking on it
would attempt to run it and it would have to be opened with a specified
editor in order to review it.