char_switch() function is fixed
char_switch() no longer switches the randomly chosen
character with a whitespace. By moving all conditionals to
a while loop, the if...else in the body of the function becomes
unnecessary as the first character in the line cannot be chosen.

Expounded upon the docstring and renamed the variables to be more
expressive of their purpose eliminating the need for unnecessary
comments.

Removed this error from the todo list.