Added full_bracket_switch()
While single_bracket_switch() replaces a single bracket type in a line
with another bracket type, full_bracket_switch() replaces all bracket
types found in a line with another bracket type. Hopefully, this will
create a logical error instead of just a syntax error.

For this to work, a new function, all_char_switch() was created.
While single_char_switch() replaces a single chosen character
with another specified character once in a line, all_char_switch()
makes those changes as many times in a line as it can.
It can take up to four characters which is necessary for
all_bracket_switch() to work.

Some functions and docstrings were edited to reflect these changes,
and the todo list was updated.

The README was not updated in this commit.