Added docstring_detect(), modified bugged_docstring()
The docstring_detect() function was created so that
a global list of all the lines which are part of a
docstring can be excluded from having bugs applied to them.
As an example, the character switching bug is meaningless
if applied to a docstring.

That full docstring detection functionality was not yet implemented
in this commit.

So far, only the functionality of generating a list of docstring
lines has been created. bugged_docstring() was changed to now use
docstring_detect() to provide the necessary lists of line indexes.

The next step is to create a global docstring list, that the
other bugstring functions can check against. If the chosen random
line index is in the global docstring list, choose another line.