Created Global Variable: docstring_line_indexes []
docstring_line_indexes is a global list which gets populated
with the lines of all the file's docstrings by the
docstring_detect() function.

The random_line() function uses it to ensure that changes aren't
made to docstrings. If a character was switched in a docstring,
it would be meaningless and wouldn't cause any errors.

NOTE: testing revealed that the elif_else_switch() function
was broken by the recent changes to single_char_switch.
It's commented out for now until a new method is used to
change multi character strings.