single_char_switch() changed to multi_char_switch()
single_char_switch has been adapted to work with words
instead of single characters. Additionally, two words
or various lengths can be supplied to it as well.

A while loop was added to make sure the line contains
at least one of the characters to be switched. However,
perhaps the file contains none of the characters and an infinite
loop would result. To resolve this, a condition was added
to ensure the while loop isn't run more than 15 times. num_bugs is
also decremented as a fail safe to ensure the whole program
doesn't enter an infinite loop.

Some bugs were detected only when running the program 500 times on the same
file. Various indexes were out of range. This needs to be addressed.