multi_char_switch() can remove functions from bug_function_list
If multi_char_switch() can't find a line with the supplied characters
more than 100 times, it will exit the function and remove it from the
global function list. This is only if the function name was supplied
to multi_char_switch(). As of this commit, only elif_else_switch()
has been setup this way. Other multi_char_switch() functions will
need to follow suit as well.