char_switch() bug fix: random_line()
Found a bug in char_switch() when running buggify() 500+ times.
random_line() had a conditional put in to ensure that
no blank lines were retrieved. However, some lines with only
one character were supplied to char_switch() and it got stuck
in an infinite loop when trying to choose a character from
the line. By changing the conditional to at least 2 unique
characters, this bug was fixed and char_switch() could be run
1500+ times.