Nerfed missing_blanks() with while loop conditional
If left unchecked, this function can be applied to the
same line multiple times and to multiple lines.
When running this program a large amount of times on the same file,
there are way too many underscores.

To combat this, it will try 20 times to find a line with less than 5 underscores
and if it can't it will remove itself from the main bug_function_list.

In this way, the function is much more reasonable.