Added hints for scrambled_line() and missing_blanks()
Comments were added to the end of these lines to give
the user hints about what needs to be done. These two
functions are slightly different than the rest in the
sense that they don't introduce bugs/errors as much
as just making the user try to deduce what the
line was supposed to say. I was originally hesitant to do
so because missing_blanks() could be applied to the same
line twice, but now that it can't, it's much less of a problem.