Rearranged entire file and added comments
While this commit appears to be a big change,
none of the code was altered. Comments were added
to split up the file into three sections and each
section was explained with an overview.
"Helper Functions" were moved to their own section
to be more clear as to what their intended use is.
All the functions in bug_function_list were put into their own section.
The global variables were put together at the bottom of the file.
The list of bug functions was rearranged to be more
coherent. It follows a more logical pattern.