Passing arguments to command line Buggify
Working on making Buggify easier to run for
end users.

Used pyinstaller to create .exe of Buggify.
pip install pyintsaller
pyinstaller .\Buggify.py --onefile

Buggify.exe worked well, but I wasn't able to
pass any arguments to it.

This commit adds the ability to provide up to two
command line arguments to Buggify in any order.
The command line arguments only apply if Buggify
is ran from command line. If it is imported and ran
from the Python shell, the original arguments apply.

In addition, the original arguments now work with up to
two arguments and in any order.

Putting this on testing branch temporarily, since it requires
more testing to ensure it works properly. As well as on
windows and linux.