Consolidation of both files into one
Edited Buggify.py to include all of buggify_functions.py
and then deleted the second file.

In the interest of making this program easier to import and run,
I thought it would be wise to consolidate both files.

Although the program is 757 lines at this point,
I think ease of use is more important.

In order to make it more readable/manageable, I put
a lot of comments explaining the subdivision and
structure of the file. I feel that it is a fair
compromise.