Fix buggify() - len(filelist)
If a file has three lines or less, it will be too small and will
encounter errors. However, if a file has three lines but is padded
with a lot of blank lines, it will pass this conditional. In order
to ensure that the file only has three lines regardless of white space,
len(filelist) was changed to len(set(filelist)) <= 4 so that the smallest
list will be three lines + 1 white space even if there are 100s of blank
lines.