Edit buggify() - test file sizes
Instead of checking the length of the file once its been
converted to a list, now there's a conditional to check
the file size of every file before we begin working on it.

If the file is too small, it's rejected. This should rule out
all the issues that arise when a file is only a few lines long.
The main bug that prompted this was a file that was more than 5
lines in length but had only one line that didn't begin with a "#".
This made random_line() results in an infinite loop.

If the file is too big, it's rejected. buggify() is only intended
to be used on text files. I suppose we could make a list of all the
accepted file types and check to make sure the file extension is
in that list, but excluding all files larger than 3mb should be an easy
way of ruling out most non-text files. The issue that prompted this
was when buggify() failed to run on an .mp4 video. Hopefully, this prevents
most binary files from being submitted and failing.