Fix Bug - Functions removing themselves
If a function searches for a line, character or word in the file
but can't find it after a certain amount of tries, then the
function will remove itself from self.bug_function_list so that
it isn't attempted to run again. After introducting the classes,
this was broken as the functions to remove no longer existed.
"equal_switch" is not defined, but self.equal_switch is.

In multi_char_switch and all_char_switch we don't need to change
func_name to self.func_name because the full function name
is already passed to it.