Deleted Buggify(.exe)
See issue #14. It's too early in this project to be worried
about producing an .exe file for every commit with updates.
In addition, pyinstaller creates an OS specific file with
regards to which OS its used on. We can delete this for now,
and we can always create it later.